+++ app/src/main/java/com/preonboarding/locationhistory/feature/history/domain/usecase/GetHistoryUseCase.kt
@@ -1,7 +1,7 @@
 package com.preonboarding.locationhistory.feature.history.domain.usecase
 
 import com.preonboarding.locationhistory.data.entity.History
-import com.preonboarding.locationhistory.domain.MapRepository
+import com.preonboarding.locationhistory.feature.history.domain.MapRepository
 import javax.inject.Inject
 
 class GetHistoryUseCase @Inject constructor(
