+++ app/src/main/java/com/preonboarding/locationhistory/feature/history/domain/usecase/SaveHistoryUseCase.kt
@@ -3,7 +3,7 @@ package com.preonboarding.locationhistory.feature.history.domain.usecase
 import com.preonboarding.locationhistory.data.entity.History
 import com.preonboarding.locationhistory.data.entity.toFormatDate
 import com.preonboarding.locationhistory.data.entity.toFormatTime
-import com.preonboarding.locationhistory.domain.MapRepository
+import com.preonboarding.locationhistory.feature.history.domain.MapRepository
 import javax.inject.Inject
 
 class SaveHistoryUseCase @Inject constructor(
